Change fps target location
