+++ Assets/Scripts/UI/MainMenu.cs
@@ -1,5 +1,6 @@
 ﻿using TMPro;
 using UnityEngine;
+using UnityEngine.Serialization;
 using UnityEngine.UI;
 
 public class MainMenu : MonoBehaviour {
@@ -8,10 +9,53 @@ public class MainMenu : MonoBehaviour {
   public TMP_InputField projectNameField;
   public Button confirmProjectButton;
   public Toggle fullscreenToggle;
+  public TMP_Dropdown vSyncRate;
+  public TMP_Dropdown fpsTarget;
 
   void Awake() {
     instance = this;
     fullscreenToggle.onValueChanged.AddListener(SetFullScreen);
+    
+    // We opt to use vSync by default with vSyncCount of 1/2
+    vSyncRate.value = PlayerPrefs.GetInt("vSyncRate", 2);
+    // By default TargetFrames is 0 (-1)
+    fpsTarget.value = PlayerPrefs.GetInt("fpsTarget", 0);
+  }
+  
+  public void SetVSyncRatio(System.Int32 value)
+  {
+    // Clear fpsTarget
+    PlayerPrefs.SetInt("fpsTarget", 0);
+    fpsTarget.value = 0;
+    Application.targetFrameRate = -1;
+		
+    if (value == 0)
+    {
+      SetFpsTarget(3);
+    }
+		
+    PlayerPrefs.SetInt("vSyncRate", value);
+    vSyncRate.value = value;
+		
+    QualitySettings.vSyncCount = value;
+  }
+
+  public void SetFpsTarget(System.Int32 value)
+  {
+    // Clear vSync Count
+    PlayerPrefs.SetInt("vSyncRate", 0);
+    vSyncRate.value = 0;
+    QualitySettings.vSyncCount = 0;
+		
+    if (value == 0)
+    {
+      SetVSyncRatio(2);
+    }
+		
+    PlayerPrefs.SetInt("fpsTarget", value);
+    fpsTarget.value = value;
+		
+    Application.targetFrameRate = value != 0 ? value * 10 : -1;
   }
 
   void LateUpdate() {
@@ -25,13 +69,6 @@ public class MainMenu : MonoBehaviour {
     string projectName = projectNameField.text;
     SaveSystem.SetActiveProject(projectName);
     UnityEngine.SceneManagement.SceneManager.LoadScene(1);
-  }
-
-  [System.Obsolete("UI/MainMenu::SetVSyncRatio is deprecated," +
-                   " please use Graphics/TargetFrameRate::SetVSyncRatio instead.")] 
-  public void SetVSyncRatio(System.Int32 value)
-  {
-    // This is implemented directly on Graphics/TargetFrameRate.cs
   }
 
   public void SetResolution16x9(int width) {
